resume/functions.py
from django.utils import timezone
from guardian.shortcuts import assign_perm, remove_perm
from django.contrib.auth.models import User
import django.forms
import django.http.request
import django.db.models


