requirements.txt
Django==2.2.6
django-appconf==1.0.3
django-braces==1.13.0
django-compressor==2.3
django-filters==0.2.1
django-guardian==2.1.0
django-libsass==0.7
django-model-utils==3.2.0
libsass==0.19.3
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.2
rcssmin==1.0.6
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
swapper==1.1.1
